pyproject.toml
[tool.poetry]
name = "groceries"
version = "0.1.0"
description = "Automatically build grocery orders."
authors = ["Nick DiRienzo <me@nickdirienzo.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.7"
beautifulsoup4 = "^4.9.1"
requests = "^2.24.0"
selenium = "^3.141.0"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
